Updated examples

+++ special_key_hash_test.c
@@ -1,5 +1,5 @@
 
-#include "hash_flex.h"
+#include "special_key_hash.h"
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -12,30 +12,30 @@ static unsigned int random_key() {
 
 int main() {
   int n = 10;
-  unsigned int keys[n];
+  size_t keys[n];
   for (int i = 0; i < n; ++i) {
     keys[i] = random_key();
   }
 
   struct hash_table *table = new_table(32);
   for (int i = 0; i < n; ++i) {
-    printf("inserting key %u\n", keys[i]);
+    printf("inserting key %lu\n", keys[i]);
     insert_key(table, keys[i]);
   }
   printf("\n");
 
   for (int i = 0; i < n; ++i) {
-    printf("is key %u in table? %d\n", keys[i], contains_key(table, keys[i]));
+    printf("is key %lu in table? %d\n", keys[i], contains_key(table, keys[i]));
   }
   printf("\n");
 
-  printf("Removing keys 3 and 4 (%u and %u)\n", keys[3], keys[4]);
+  printf("Removing keys 3 and 4 (%lu and %lu)\n", keys[3], keys[4]);
   delete_key(table, keys[3]);
   delete_key(table, keys[4]);
   printf("\n");
 
   for (int i = 0; i < n; ++i) {
-    printf("is key %u in table? %d\n", keys[i], contains_key(table, keys[i]));
+    printf("is key %lu in table? %d\n", keys[i], contains_key(table, keys[i]));
   }
   printf("\n");
